feat(pm): implement PM credential API with C permission semantics
Port pm/getset.c do_set/do_get and wire the VFS forward path, so
non-root processes can no longer escalate and no credential call
routes to no_sys.

- pm.rs do_set: EPERM gates per C (setuid/seteuid/setgid/setegid
  allowed only when new id equals the real id, or by root);
  SETUID/SETGID update real+effective, SETEUID/SETEGID effective only
- pm.rs do_setgroups: root-only, NGROUPS_MAX/EINVAL checks, gids above
  0xFFFF rejected, caller's group list copied via sys_vircopy
- New dispatch entries: PM_SETGROUPS(9), PM_GETGROUPS(10),
  PM_SETEUID(29), PM_SETEGID(30), PM_ISSETUGID(31), PM_GETSID(32),
  matching C callnr.h
- pm.rs do_newexec: setuid exec keeps the real uid (effective only);
  TAINTED now follows C (cleared first, set only on setuid/setgid exec
  or real!=eff mismatch)
- vfs/pm.rs: VFS_PM_SETGROUPS service arm copies PM's i32 gids into a
  NGROUPS_MAX buffer and stores them via pm_setgroups, replies
  VFS_PM_RS_BASE+11
- vfs/misc.rs: pm_setgroups takes *const i32 (was *const u16)

Verified: servers 575 host tests (10 new, fail-pre-change proven),
clippy -D warnings clean, cargo test --workspace green,
test-boot-x86 ALL TESTS PASSED, exec loop flat (0 regions/exec).

+++ crates/servers/src/vfs/misc.rs
@@ -227,17 +227,21 @@ pub fn pm_setuid(proc_e: i32, euid: i32, ruid: i32) {
 
 /// Set groups for a process.
 ///
+/// `addr` points to an array of `ngroups` i32 gids in the caller's (PM's)
+/// address space — the service_pm SETGROUPS arm copies it here first.
+///
 /// # Safety
 ///
-/// `addr` must point to a valid array of at least `ngroups` u16 values.
-pub unsafe fn pm_setgroups(proc_e: i32, ngroups: i32, addr: *const u16) {
+/// `addr` must point to a valid array of at least `ngroups` i32 values
+/// when `ngroups > 0`.
+pub unsafe fn pm_setgroups(proc_e: i32, ngroups: i32, addr: *const i32) {
     if let Some(rfp) = fproc_by_endpoint(proc_e) {
         unsafe {
             (*rfp).fp_ngroups = ngroups;
             if ngroups > 0 && !addr.is_null() {
                 let count = ngroups.min(NGROUPS_MAX as i32) as usize;
                 for i in 0..count {
-                    (*rfp).fp_sgroups[i] = *addr.add(i);
+                    (*rfp).fp_sgroups[i] = (*addr.add(i)) as u16;
                 }
             }
         }
@@ -508,7 +512,7 @@ mod tests {
         unsafe {
             (*fp_slot(2)).fp_endpoint = 2;
         }
-        let groups = [100u16, 200, 300];
+        let groups = [100i32, 200, 300];
         unsafe {
             pm_setgroups(2, 3, groups.as_ptr());
         }
